# backend/src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: MadaRisk Map - API Backend
  description: |
    Plateforme SIG de surveillance et d'anticipation des catastrophes naturelles à Madagascar.

    La plupart des endpoints nécessitent un jeton d'accès JWT.
    Obtenez-le via `POST /auth/login`, puis cliquez sur **Authorize** et renseignez `Bearer <votre_token>`.

    Conventions :
    - Réponse de succès : `{ success: true, message, data, meta? }`.
    - Réponse d'erreur : `{ success: false, message, errors? }`.
    - Pagination : `page` (défaut 1), `limit` (défaut 20, max 100), `meta: { page, limit, total, totalPages }`.
    - Erreurs : `400` requête invalide, `401` non authentifié, `403` rôle insuffisant,
      `404` introuvable, `409` conflit, `422` échec de validation (zod), `429` trop de requêtes,
      `500` erreur interne.
    - Rôles : `SUPER_ADMIN`, `ADMIN`, `ANALYSTE_SIG`, `CLIENT`.
  version: 1.0.0
servers:
  - url: http://localhost:5000/api/v1
    description: Serveur local de développement - API versionnée
  - url: http://localhost:5000
    description: Serveur local de développement - racine (health + documentation)
...
paths:
...
  /events/{id}/areas/calculate:
    post:
      tags: [Événements]
      summary: Calculer une zone d'influence (ADMIN / SUPER_ADMIN)
      description: Crée un cercle d'influence de rayon `radiusKm` (1-500) autour du dernier point de trajectoire, pour une phase et un niveau de risque.
      security:
        - bearerAuth: []
      operationId: calculateArea
      parameters:
        - { $ref: '#/components/parameters/EventId' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phase, riskLevel, radiusKm]
              properties:
                phase: { $ref: '#/components/schemas/RiskPhase' }
                riskLevel: { $ref: '#/components/schemas/RiskLevel' }
                radiusKm:
                  type: number
                  minimum: 1
                  maximum: 500
                  description: Rayon de la zone d'influence en km.
            examples:
              influenceAreaExample:
                $ref: '#/components/examples/influenceArea'
      responses:
        '201':
          description: Zone d'influence calculée
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  message: { type: string }
                  data: { $ref: '#/components/schemas/TrackInfluenceArea' }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
  /events/{id}/exposure/calculate:
    post:
      tags: [Événements]
      summary: Calculer l'exposition des communes (ADMIN / SUPER_ADMIN)
      description: Détermine les communes intersectant la zone d'influence (population estimée, superficie touchée, enjeux).
      security:
        - bearerAuth: []
      operationId: calculateExposure
      parameters:
        - { $ref: '#/components/parameters/EventId' }
        - { name: areaId, in: query, required: false, schema: { type: string, format: uuid }, description: Zone d'influence ciblée }
        - { name: allAreas, in: query, required: false, schema: { type: boolean }, description: Utiliser toutes les zones ("true" pour oui) }
      responses:
        '200':
          description: Exposition calculée
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  message: { type: string }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/ExposureResult' }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Jeton d'accès obtenu via `POST /auth/login`.
  parameters:
...
    EventId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    BadRequest:
      description: Requête invalide
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorBody' }
    Unauthorized:
      description: Non authentifié (jeton manquant, invalide ou expiré)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorBody' }
    Forbidden:
      description: Rôle insuffisant (opérations réservées aux SUPER_ADMIN / ADMIN / ANALYSTE_SIG)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorBody' }
    NotFound:
      description: Ressource introuvable
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorBody' }
...
    ValidationError:
      description: Échec de validation (zod) - détails champ par champ
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ValidationBody' }
...
    InternalError:
      description: Erreur interne du serveur
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorBody' }
  examples:
...
    influenceArea:
      summary: Calcul d'une zone d'influence
      value:
        phase: PENDANT
        riskLevel: ELEVE
        radiusKm: 75
...
  schemas:
    ErrorBody:
      type: object
      description: Corps d'erreur standard
      properties:
        success:
          type: boolean
          enum: [false]
        message: { type: string }
        errors:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              message: { type: string }
    ValidationBody:
      type: object
      description: Erreur de validation (422)
      properties:
        success:
          type: boolean
          enum: [false]
        message: { type: string }
        errors:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              message: { type: string }
...
    RiskLevel:
      type: string
      enum: [FAIBLE, MODERE, ELEVE, EXTREME]
    RiskPhase:
      type: string
      enum: [AVANT, PENDANT, APRES, RETABLISSEMENT]
...
    GeoJsonGeometry:
      type: object
      description: Géométrie GeoJSON (Point, LineString, Polygon, MultiPolygon, cercle)
      properties:
        type:
          type: string
          enum: [Point, LineString, Polygon, MultiPolygon, Circle]
        coordinates: {}
        radius: { type: number, description: Rayon en km pour une géométrie de type Circle }
...
    TrackInfluenceArea:
      type: object
      description: Zone d'influence calculée autour d'un point de trajectoire
      properties:
        id: { type: string, format: uuid }
        eventId: { type: string, format: uuid }
        trackId: { type: string, format: uuid }
        phase: { $ref: '#/components/schemas/RiskPhase' }
        radiusKm: { type: number }
        riskLevel: { $ref: '#/components/schemas/RiskLevel' }
        circle: { $ref: '#/components/schemas/GeoJsonGeometry' }
        createdAt: { type: string, format: date-time }
    ExposureResult:
      type: object
      description: Résultat du calcul d'exposition d'une commune
      properties:
        communeId: { type: string, format: uuid }
        communeName: { type: string }
        districtId: { type: string, format: uuid }
        distanceToTrackKm: { type: number }
        areaRadiusKm: { type: number }
        phase: { $ref: '#/components/schemas/RiskPhase' }
        populationEstimated: { type: integer }
        householdsEstimated: { type: integer }
        areaKmTouched: { type: number }
...